Create an enum for reading errors
When a sensor reads a temperature we might encounter an error, if
the file is not available or if no sensor is detected for example.

That's why we need an error type to know what's the problem during a
read.

An enumeration has been created to match the possible errors we might
encounter during the reading.