Create a configurations module
The managers will use the items in this module to read the configuration
values.

The ConfigurationReader trait aims to standardize the configuration
reading. Doing so a user would be able to use different configuration
format like .env, YAML or XML. The managers would not need to care about
the configuration format to implement.