The get_value's return type must implement TryFrom
A configuration reader must implement a method "get_value" that return
a value from the configuration. These datas may be converted into a
custom type since the method is generic.

Since the values will be read as string from a any source, the return
type of "get_value" must implement the TryFrom trait. Doing so a value
could used to instantiate a custom structure or enumeration.