Create a dot_env configuration reader
The dot_env configuration reader is not finished yet, but it works.
First it reads a .env file and puts the variables in the environment,
then it's able to get the values from the environment.

In the future it will store the values in an attribute so it's not
concerned by any environement modification. The .env file would be only
read as a format, but will not be used to store datas into environnement.