Temperature becomes Libgrowbloom
Most of the elements defined here was quite common to many projects
GrowBloom could develop. In fact, except the "Temperature" struct
everything was generic.

The Temperature structure was removed, and the ExportedData structure
has been updated to be generic, so it's able to export any kind of data
according to the Exporter's implementation.

The temperature things will come back in a dedicated program, but this
one is dedicated to the common things GrowBloom could use in many of
its programs.