Configuration reader's get_value convert the value
A configuration reader has a "get_value" method used to return a value
from the configuration. This method returned a String because
configuration is usually written under the form of key => value pairs.

Since a configuration may be more complex, contains a whole database
configuration under a "database" key in YAML for example, it may be
useful to let the configuration reader convert its configuration value to
the given type.

The "get_value" is now generic. It returns a type that implements
TryFrom<String>, so any item implementing this trait could be returned
by the method.