The exported item has to implement Converted
An Exporter is generic and has two types, T and U. T is the type
contained by the ExportedData, and U is the type T should be
able to convert to.
In brief, T must implement Convertable<U>.

Doing so the Stdout exporter is sure the models it has to export
may be converted to String before printing.